backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==2.3.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
PuLP==3.2.1
Werkzeug==3.1.3
